cloud-run/filtering/requirements.txt
astropy
google-cloud-functions
google-cloud-logging
pittgoogle-client>=0.3.12

# for Cloud Run
# https://cloud.google.com/run/docs/quickstarts/build-and-deploy/deploy-python-service
# pinned following quickstart example. [TODO] consider un-pinning
Flask
gunicorn
Werkzeug
